Todos los modulos listos

+++ src/main/java/com/Proyecto/ONU/Controllers/CiudadController.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Controllers;
 
-import java.math.BigInteger;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -74,7 +74,7 @@ public class CiudadController {
     
 
     @PutMapping("/{id}")
-    public ResponseEntity<Map<String, Object>> update(@PathVariable BigInteger id, @Validated @RequestBody CiudadDTO ciudadDTO, BindingResult result) {
+    public ResponseEntity<Map<String, Object>> update(@PathVariable Long id, @Validated @RequestBody CiudadDTO ciudadDTO, BindingResult result) {
 
         Map<String, Object> response = new HashMap<>();
 
@@ -106,7 +106,7 @@ public class CiudadController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable BigInteger id){
+    public void deleteById(@PathVariable Long id){
         ciudadService.deleteById(id);
     } 
 }
