+++ src/main/java/com/Proyecto/ONU/Controllers/PersonaController.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Controllers;
 
-import java.math.BigInteger;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -72,7 +72,7 @@ public class PersonaController {
     
 
     @PutMapping("/{id}")
-    public ResponseEntity<Map<String, Object>> update(@PathVariable BigInteger id, @Validated @RequestBody PersonaDTO personaDTO, BindingResult result) {
+    public ResponseEntity<Map<String, Object>> update(@PathVariable Long id, @Validated @RequestBody PersonaDTO personaDTO, BindingResult result) {
 
         Map<String, Object> response = new HashMap<>();
 
@@ -104,7 +104,7 @@ public class PersonaController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable BigInteger id){
+    public void deleteById(@PathVariable Long id){
         personaService.deleteById(id);
     } 
 }
