+++ src/main/java/com/Proyecto/ONU/Controllers/TipoCuotaController.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Controllers;
 
-import java.math.BigInteger;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -71,7 +71,7 @@ public class TipoCuotaController {
     
 
     @PutMapping("/{id}")
-    public ResponseEntity<Map<String, Object>> update(@PathVariable BigInteger id, @Validated @RequestBody TipoCuotaDTO tipoCuotaDTO, BindingResult result) {
+    public ResponseEntity<Map<String, Object>> update(@PathVariable Long id, @Validated @RequestBody TipoCuotaDTO tipoCuotaDTO, BindingResult result) {
 
         Map<String, Object> response = new HashMap<>();
 
@@ -103,7 +103,7 @@ public class TipoCuotaController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable BigInteger id){
+    public void deleteById(@PathVariable Long id){
         tipoCuotaService.deleteById(id);
     } 
 }
