+++ src/main/java/com/Proyecto/ONU/Repository/CuotaRepository.java
@@ -1,11 +1,11 @@
 package com.Proyecto.ONU.Repository;
 
-import java.math.BigInteger;
+
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.Proyecto.ONU.Repository.Entities.Cuota;
 
-public interface CuotaRepository extends JpaRepository<Cuota, BigInteger>{
+public interface CuotaRepository extends JpaRepository<Cuota, Long>{
     
 }
