+++ src/main/java/com/Proyecto/ONU/Repository/EntitiesDTO/CiudadDTO.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Repository.EntitiesDTO;
 
-import java.math.BigInteger;
+
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 public class CiudadDTO {
 
-    private BigInteger id;
+    private Long id;
 
     @NotEmpty(message = "no puede estar vacio")
     private String nombre;
