+++ src/main/java/com/Proyecto/ONU/Repository/MaterialRepository.java
@@ -1,12 +1,12 @@
 package com.Proyecto.ONU.Repository;
 
-import java.math.BigInteger;
+
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.Proyecto.ONU.Repository.Entities.Material;
 
-public interface MaterialRepository extends JpaRepository<Material,BigInteger>{
+public interface MaterialRepository extends JpaRepository<Material,Long>{
 
 
 }
