+++ src/main/java/com/Proyecto/ONU/Repository/PersonaRepository.java
@@ -1,11 +1,11 @@
 package com.Proyecto.ONU.Repository;
 
-import java.math.BigInteger;
+
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.Proyecto.ONU.Repository.Entities.Persona;
 
-public interface PersonaRepository extends JpaRepository<Persona, BigInteger> {
+public interface PersonaRepository extends JpaRepository<Persona, Long> {
     
 }
