+++ src/main/java/com/Proyecto/ONU/Repository/UsuarioRepository.java
@@ -1,12 +1,12 @@
 package com.Proyecto.ONU.Repository;
 
-import java.math.BigInteger;
+
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.Proyecto.ONU.Repository.Entities.Usuario;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, BigInteger>{
+public interface UsuarioRepository extends JpaRepository<Usuario, Long>{
     Optional<Usuario> findByEmail(String email);
 }
