+++ src/main/java/com/Proyecto/ONU/Service/CiudadService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.CiudadDTO;
@@ -10,9 +10,9 @@ public interface CiudadService {
 
     public CiudadDTO save(CiudadDTO ciudadDTO);
 
-    public CiudadDTO update(BigInteger id, CiudadDTO ciudadDTO );
+    public CiudadDTO update(Long id, CiudadDTO ciudadDTO );
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<CiudadDTO> findAll();
 }
