+++ src/main/java/com/Proyecto/ONU/Service/CuotaService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.CuotaDTO;
@@ -9,9 +9,9 @@ public interface CuotaService {
 
     public CuotaDTO save(CuotaDTO cuotaDTO );
 
-    public CuotaDTO  update(BigInteger id, CuotaDTO cuotaDTO );
+    public CuotaDTO  update(Long id, CuotaDTO cuotaDTO );
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<CuotaDTO > findAll();
