+++ src/main/java/com/Proyecto/ONU/Service/EnvioMaterialService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.EnvioMaterialDTO;
@@ -9,11 +9,11 @@ public interface EnvioMaterialService {
 
     public EnvioMaterialDTO save(EnvioMaterialDTO envioMaterialDTO);
 
-    public EnvioMaterialDTO update(BigInteger id, EnvioMaterialDTO envioMaterialDTO);
+    public EnvioMaterialDTO update(Long id, EnvioMaterialDTO envioMaterialDTO);
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<EnvioMaterialDTO> findAll();
 
-    EnvioMaterialDTO findById(BigInteger id);
+    EnvioMaterialDTO findById(Long id);
 }
