+++ src/main/java/com/Proyecto/ONU/Service/MaterialService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.MaterialDTO;
@@ -9,9 +9,9 @@ public interface MaterialService {
 
     public MaterialDTO save(MaterialDTO materialDTO);
 
-    public MaterialDTO  update(BigInteger id, MaterialDTO materialDTO);
+    public MaterialDTO  update(Long id, MaterialDTO materialDTO);
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<MaterialDTO> findAll();
 }
