+++ src/main/java/com/Proyecto/ONU/Service/PersonaService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.PersonaDTO;
@@ -9,9 +9,9 @@ public interface PersonaService {
     
     public PersonaDTO save(PersonaDTO personaDTO);
 
-    public PersonaDTO update(BigInteger id, PersonaDTO personaDTO);
+    public PersonaDTO update(Long id, PersonaDTO personaDTO);
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<PersonaDTO> findAll();
