+++ src/main/java/com/Proyecto/ONU/Service/RefugioService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.RefugioDTO;
@@ -9,9 +9,9 @@ public interface RefugioService {
     
     public RefugioDTO save(RefugioDTO refugioDTO);
 
-    public RefugioDTO update(BigInteger id, RefugioDTO refugioDTO);
+    public RefugioDTO update(Long id, RefugioDTO refugioDTO);
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<RefugioDTO> findAll();
 }
