+++ src/main/java/com/Proyecto/ONU/Service/SedeService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.SedeDTO;
@@ -10,9 +10,9 @@ public interface SedeService {
 
     public SedeDTO save(SedeDTO sedeDTO);
 
-    public SedeDTO update(BigInteger id, SedeDTO sedeDTO);
+    public SedeDTO update(Long id, SedeDTO sedeDTO);
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<SedeDTO> findAll();
 }
