+++ src/main/java/com/Proyecto/ONU/Service/SocioService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.SocioDTO;
@@ -9,13 +9,13 @@ public interface SocioService {
 
     public SocioDTO save(SocioDTO socioDTO);
 
-    public SocioDTO update(BigInteger id, SocioDTO persona);
+    public SocioDTO update(Long id, SocioDTO persona);
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<SocioDTO> findAll();
 
-    List<SocioDTO> findAllSocioByTipoCuota(BigInteger ipoCuotaid);
+    List<SocioDTO> findAllSocioByTipoCuota(Long ipoCuotaid);
     
     List<SocioDTO> InformeCuotaPorSocio();
 }
