+++ src/main/java/com/Proyecto/ONU/Service/TipoCuotaService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.TipoCuotaDTO;
@@ -9,9 +9,9 @@ public interface TipoCuotaService {
     
     public TipoCuotaDTO save(TipoCuotaDTO tipoCuotaDTO);
 
-    public TipoCuotaDTO update(BigInteger id, TipoCuotaDTO tipoCuotaDTO);
+    public TipoCuotaDTO update(Long id, TipoCuotaDTO tipoCuotaDTO);
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<TipoCuotaDTO> findAll();
 }
