+++ src/main/java/com/Proyecto/ONU/Service/UsuarioService.java
@@ -1,6 +1,6 @@
 package com.Proyecto.ONU.Service;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 import com.Proyecto.ONU.Repository.EntitiesDTO.UsuarioDTO;
@@ -11,9 +11,9 @@ public interface UsuarioService {
 
     public UsuarioDTO save(UsuarioDTO usuario);
 
-    public UsuarioDTO update(BigInteger id, UsuarioDTO usuario);
+    public UsuarioDTO update(Long id, UsuarioDTO usuario);
 
-    void deleteById(BigInteger id);
+    void deleteById(Long id);
 
     List<UsuarioDTO> findAll();
